Move align, map, abundance modes to separate module
They are now submodules of a new 'modes' module.
