+++ src/lib.rs
@@ -8,10 +8,9 @@ pub mod index;
 pub mod indexer;
 pub mod insertsize;
 pub mod io;
-pub mod maponly;
-pub mod mapper;
 pub mod math;
 pub mod mcsstrategy;
+pub mod modes;
 pub mod packed_seq;
 pub mod partition;
 pub mod piecewisealigner;
